+++ Sources/Models/IM.swift
@@ -12,7 +12,7 @@ public struct IM {
     public let is_open: Bool
     public let has_pins: Bool
     public let user: User
-    public let created: Double
+    public let created: Int
     public let is_user_deleted: Bool
     public let last_read: String
     public let latest: Message?
